Поправила нерабовшие тесты
